add table for user management
