+++ client/src/components/nav-bar/nav-bar.tsx
@@ -1,0 +1,12 @@
+
+
+import React, { useEffect, useState } from 'react'
+export const NavBar = () => {
+	console.log('list table')
+
+	return (
+		<div>
+            Menu Bar
+		</div>
+	)
+}
