+++ client/src/pages/landing-page/index.tsx
@@ -1,0 +1,15 @@
+import type { NextPage } from 'next';
+import React, { useEffect, useState } from 'react';
+import { NavBar } from '../../components/nav-bar/nav-bar';
+
+const LadingPage: NextPage = () => {
+  return (
+    <div className="flex min-h-screen flex-col items-center justify-center py-2">
+      <NavBar />
+      <div>
+        
+      </div>
+    </div>
+  )
+}
+export default LadingPage
